proj/RecommendationService/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic>=2.10.0
pydantic-settings==2.12.0
python-dotenv==1.0.0
qdrant_client==1.16.0
numpy==2.3.5
psycopg2-binary>=2.9.0
redis>=5.0.0
